SONARGO-113 Rewrite tests of SwitchWithoutDefaultCheck
