+++ sonar-go-checks/src/main/java/org/sonar/go/checks/GoCheckList.java
@@ -48,7 +48,7 @@ public class GoCheckList {
       IdenticalConditionsCheck.class,
       IfConditionalAlwaysTrueOrFalseCheck.class,
       SwitchCaseTooBigCheck.class,
-      MatchWithoutElseCheck.class,
+      SwitchWithoutDefaultCheck.class,
       NestedMatchCheck.class,
       OctalValuesCheck.class,
       OneStatementPerLineGoCheck.class,
