+++ sonar-go-checks/src/test/java/org/sonar/go/checks/SwitchWithoutDefaultCheckTest.java
@@ -16,24 +16,13 @@
  */
 package org.sonar.go.checks;
 
-import org.sonar.check.Rule;
-import org.sonarsource.slang.api.MatchTree;
-import org.sonarsource.slang.api.Token;
-import org.sonarsource.slang.checks.api.InitContext;
-import org.sonarsource.slang.checks.api.SlangCheck;
+import org.junit.jupiter.api.Test;
 
-@Rule(key = "S131")
-public class MatchWithoutElseCheck implements SlangCheck {
+class SwitchWithoutDefaultCheckTest {
 
-  @Override
-  public void initialize(InitContext init) {
-    init.register(MatchTree.class, (ctx, tree) -> {
-      if (tree.cases().stream().noneMatch(matchCase -> matchCase.expression() == null)) {
-        Token keyword = tree.keyword();
-        String message = String.format("Add a default clause to this \"%s\" statement.", keyword.text());
-        ctx.reportIssue(keyword, message);
-      }
-    });
+  @Test
+  void shouldRaiseIssues() {
+    GoVerifier.verify("SwitchWithoutDefaultCheck/SwitchWithoutDefaultCheck.go", new SwitchWithoutDefaultCheck());
   }
 
 }
